Ported Clock.cpp to the autopilot
